Test of Mean and Covariance Matrix at the Same Time

Hypothesis testing of multiple multivariate normal population mean vectors and covariance matrices (Test simultaneously).

meancov.Test(X, label, alpha = 0.05, full = FALSE)

Arguments

X

The data matrix which is a matrix or data frame.

label

A vector of group labels.

alpha

The significance level. Default is 0.05.

full

A boolean value. Default is FALSE. If TRUE, the full output will be displayed.

Value

If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, statistics, modify factor, df of Chi2, p value, critical value and conclusion will be returned.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.90-91.

Examples

data(iris)
chart <- iris[, 1:4]
species <- iris[, 5]
meancov.Test(chart, species)
#> $Hypothesis
#> [1] "H0: mu1 = mu2 = ... = muk, Sigma1 = Sigma2 = ... = Sigmak, k =  3"
#> 
#> $Statistics
#>                    Statistics         Value p.value Critical.Value
#> 1            Likelihood Ratio 1.767862e-155      NA             NA
#> 2 Likelihood Ratio (Modified) 2.200389e-152      NA             NA
#> 3                           M  6.984086e+02      NA             NA
#> 4                        Chi2  6.809880e+02       0       41.33714
#> 
#> $Conclusion
#> [1] "Reject"
#> 
meancov.Test(chart, species, full = TRUE)
#> $Hypothesis
#> [1] "H0: mu1 = mu2 = ... = muk, Sigma1 = Sigma2 = ... = Sigmak, k =  3"
#> 
#> $`Sample Size`
#>     setosa versicolor  virginica 
#>         50         50         50 
#> 
#> $`Total Sample Mean`
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>     5.843333     3.057333     3.758000     1.199333 
#> 
#> $`Within Group Mean`
#> $`Within Group Mean`$setosa
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.006        3.428        1.462        0.246 
#> 
#> $`Within Group Mean`$versicolor
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.936        2.770        4.260        1.326 
#> 
#> $`Within Group Mean`$virginica
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        6.588        2.974        5.552        2.026 
#> 
#> 
#> $`Total Sum of Squares`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   102.168333   -6.322667     189.8730    76.92433
#> Sepal.Width     -6.322667   28.306933     -49.1188   -18.12427
#> Petal.Length   189.873000  -49.118800     464.3254   193.04580
#> Petal.Width     76.924333  -18.124267     193.0458    86.56993
#> 
#> $`Within Sum of Squares`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      38.9562     13.6300      24.6246      5.6450
#> Sepal.Width       13.6300     16.9620       8.1208      4.8084
#> Petal.Length      24.6246      8.1208      27.2226      6.2718
#> Petal.Width        5.6450      4.8084       6.2718      6.1566
#> 
#> $Statistics
#>                    Statistics         Value p.value Critical.Value
#> 1            Likelihood Ratio 1.767862e-155      NA             NA
#> 2 Likelihood Ratio (Modified) 2.200389e-152      NA             NA
#> 3                           M  6.984086e+02      NA             NA
#> 4                        Chi2  6.809880e+02       0       41.33714
#> 
#> $`Modify Factor`
#> [1] 0.02494331
#> 
#> $`Degree of freedom`
#> [1] "Chi2( 28 )"
#> 
#> $Conclusion
#> [1] "Reject"
#>