Test of Multiple Covariance Matrix

Hypothesis testing of multiple multivariate normal population covariance matrices when the population mean is unknown.

Usage

covTest.multi(X, label, alpha = 0.05, full = FALSE)

Arguments

X

The data matrix which is a matrix or data frame.

label

A vector of group labels.

alpha

The significance level. Default is 0.05.

full

A boolean value. Default is FALSE. If TRUE, the full output will be displayed.

Value

If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample size, total sample mean, within group mean, within group sample covariance, total sum of squares, within sum of squares, statistics, modify factor, df of Chi2, p value, critical value and conclusion will be returned.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.88-89.

Author

Xifeng Zhang

Examples

data(iris)
chart <- iris[, 1:4]
species <- iris[, 5]
covTest.multi(chart, species)
#> $Hypothesis
#> [1] "H0: Sigma1 = Sigma2 = ... = Sigmak, k =  3"
#> 
#> $Statistics
#>                  Statistics        Value p.value Critical.Value
#> 1          Likelihood Ratio 3.180759e-33      NA             NA
#> 2 Modified Likelihood Ratio 1.420627e-32      NA             NA
#> 3                         M 1.466632e+02      NA             NA
#> 4                      Chi2 1.409430e+02       0       31.41043
#> 
#> $Conclusion
#> [1] "Reject"
#> 
covTest.multi(chart, species, full = TRUE)
#> $Hypothesis
#> [1] "H0: Sigma1 = Sigma2 = ... = Sigmak, k =  3"
#> 
#> $`Sample Size`
#>     setosa versicolor  virginica 
#>         50         50         50 
#> 
#> $`Total Sample Mean`
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>     5.843333     3.057333     3.758000     1.199333 
#> 
#> $`Within Group Mean`
#> $`Within Group Mean`$setosa
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.006        3.428        1.462        0.246 
#> 
#> $`Within Group Mean`$versicolor
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.936        2.770        4.260        1.326 
#> 
#> $`Within Group Mean`$virginica
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        6.588        2.974        5.552        2.026 
#> 
#> 
#> $`Within Group Sample Covariance`
#> $`Within Group Sample Covariance`$setosa
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.12424898 0.099216327  0.016355102 0.010330612
#> Sepal.Width    0.09921633 0.143689796  0.011697959 0.009297959
#> Petal.Length   0.01635510 0.011697959  0.030159184 0.006069388
#> Petal.Width    0.01033061 0.009297959  0.006069388 0.011106122
#> 
#> $`Within Group Sample Covariance`$versicolor
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.26643265  0.08518367   0.18289796  0.05577959
#> Sepal.Width    0.08518367  0.09846939   0.08265306  0.04120408
#> Petal.Length   0.18289796  0.08265306   0.22081633  0.07310204
#> Petal.Width    0.05577959  0.04120408   0.07310204  0.03910612
#> 
#> $`Within Group Sample Covariance`$virginica
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   0.40434286  0.09376327   0.30328980  0.04909388
#> Sepal.Width    0.09376327  0.10400408   0.07137959  0.04762857
#> Petal.Length   0.30328980  0.07137959   0.30458776  0.04882449
#> Petal.Width    0.04909388  0.04762857   0.04882449  0.07543265
#> 
#> 
#> $`Total Sum of Squares`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      38.9562     13.6300      24.6246      5.6450
#> Sepal.Width       13.6300     16.9620       8.1208      4.8084
#> Petal.Length      24.6246      8.1208      27.2226      6.2718
#> Petal.Width        5.6450      4.8084       6.2718      6.1566
#> 
#> $`Within Sum of Squares`
#> $`Within Sum of Squares`$setosa
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length       6.0882      4.8616       0.8014      0.5062
#> Sepal.Width        4.8616      7.0408       0.5732      0.4556
#> Petal.Length       0.8014      0.5732       1.4778      0.2974
#> Petal.Width        0.5062      0.4556       0.2974      0.5442
#> 
#> $`Within Sum of Squares`$versicolor
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      13.0552       4.174        8.962      2.7332
#> Sepal.Width        4.1740       4.825        4.050      2.0190
#> Petal.Length       8.9620       4.050       10.820      3.5820
#> Petal.Width        2.7332       2.019        3.582      1.9162
#> 
#> $`Within Sum of Squares`$virginica
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      19.8128      4.5944      14.8612      2.4056
#> Sepal.Width        4.5944      5.0962       3.4976      2.3338
#> Petal.Length      14.8612      3.4976      14.9248      2.3924
#> Petal.Width        2.4056      2.3338       2.3924      3.6962
#> 
#> 
#> $Statistics
#>                  Statistics        Value p.value Critical.Value
#> 1          Likelihood Ratio 3.180759e-33      NA             NA
#> 2 Modified Likelihood Ratio 1.420627e-32      NA             NA
#> 3                         M 1.466632e+02      NA             NA
#> 4                      Chi2 1.409430e+02       0       31.41043
#> 
#> $`Modify Factor`
#> [1] 0.03900227
#> 
#> $`Degree of freedom`
#> [1] "Chi2( 20 )"
#> 
#> $Conclusion
#> [1] "Reject"
#>