Test of Single Covariance Matrix

Hypothesis testing of covariance matrices for a single normal population when the mean vector is unknown.

Usage

covTest.single(data, Sigma0, ball = FALSE, alpha = 0.05, full = FALSE)

Arguments

data

The data matrix which is a matrix or data frame.

Sigma0

The covariance matrix when the null hypothesis is true.

ball

A boolean value. Default is FALSE. If FALSE, the covariance matrix is Sigma0 (known). If TRUE and the Sigma0 unit matrix, the Mauchly ball test is performed. If TRUE but Sigma0 is not a unit matrix, the covariance array is tested to see if it is sigma^2*Sigma0 (sigma^2 is unknown).

alpha

The significance level. Default is 0.05.

full

A boolean value. Default is FALSE. If TRUE, the full output will be displayed.

Value

A list containing the hypothesis, sample mean, sample diviation, statistics, df of Chi2, p value, critical value and conclusion.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.83-88.

Author

Xifeng Zhang

Examples

data(iris)
X <- iris[, 1:4]
covTest.single(X, diag(1, 4))
#> $Hypothesis
#> [1] "H0: Sigma = Sigma0"
#> 
#> $Statistics
#>         Statistics         Value p.value Critical.Value
#> 1 Likelihood Ratio 3.829315e-223      NA             NA
#> 2             Chi2  1.024268e+03       0       18.30704
#> 
#> $Conclusion
#> [1] "Reject"
#> 
covTest.single(X, diag(1, 4), ball = TRUE)
#> $Hypothesis
#> [1] "Mauchly's test of sphericity. H0: Sigma = sigma2 * Ip"
#> 
#> $Statistics
#>         Statistics         Value p.value Critical.Value
#> 1 Likelihood Ratio 8.616505e-321      NA             NA
#> 2                W  5.400950e-05      NA             NA
#> 3             Chi2  1.448568e+03       0       16.91898
#> 
#> $Conclusion
#> [1] "Reject"
#> 
covTest.single(X, diag(2, 4), ball = TRUE)
#> $Hypothesis
#> [1] NA
#> 
#> $Statistics
#>         Statistics         Value p.value Critical.Value
#> 1 Likelihood Ratio 8.616505e-321      NA             NA
#> 2                W  5.400950e-05      NA             NA
#> 3             Chi2  1.448568e+03       0       16.91898
#> 
#> $Conclusion
#> [1] "Reject"
#> 
covTest.single(X, diag(1, 4), full = TRUE)
#> $Hypothesis
#> [1] "H0: Sigma = Sigma0"
#> 
#> $`Sample Mean`
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>       5.8433       3.0573       3.7580       1.1993 
#> 
#> $`Sample Diviation`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   102.168333   -6.322667     189.8730    76.92433
#> Sepal.Width     -6.322667   28.306933     -49.1188   -18.12427
#> Petal.Length   189.873000  -49.118800     464.3254   193.04580
#> Petal.Width     76.924333  -18.124267     193.0458    86.56993
#> 
#> $Statistics
#>         Statistics         Value p.value Critical.Value
#> 1 Likelihood Ratio 3.829315e-223      NA             NA
#> 2             Chi2  1.024268e+03       0       18.30704
#> 
#> $`Degree of freedom`
#> [1] "Chi2( 10 )"
#> 
#> $Conclusion
#> [1] "Reject"
#>