Test of Single Mean Vector

Hypothesis testing of the mean vector of a single multivariate normal population when the covariance matrix is known or unknown.

Usage

meanTest.single(data, mu0, Sigma0 = FALSE, alpha = 0.05, full = FALSE)

Arguments

data

The data matrix which is a matrix or data frame.

mu0

The mean vector when the null hypothesis is true.

Sigma0

(optional) The population covariance matrix. Default is FALSE which means the covariance matrix is unknown.

alpha

The significance level. Default is 0.05.

full

A boolean value. Default is FALSE. If TRUE, the full output will be displayed.

Value

If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample mean, sample diviation, statistics, degree of freedom, p value, critical value and conclusion will be returned.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.66-68.

Author

Xifeng Zhang

Examples

data(iris)
X <- iris[, 1:4]
mu0 <- c(5.8, 3.0, 4.3, 1.3)
meanTest.single(X, mu0)
#> $Hypothesis
#> [1] "H0: mu = ( 5.8, 3, 4.3, 1.3 ) when Sigma is unknown"
#> 
#> $Statistics
#>     Statistics     Value p.value Critical.Value
#> 1 Hotelling T2 244.16277      NA             NA
#> 2            F  59.81168       0       2.433633
#> 
#> $Conclusion
#> [1] "Reject"
#> 
meanTest.single(X, mu0, Sigma0 = diag(1, 4))
#> $Hypothesis
#> [1] "H0: mu = ( 5.8, 3, 4.3, 1.3 ) when Sigma0 is known"
#> 
#> $Statistics
#>   Statistics   Value     p.value Critical.Value
#> 1       Chi2 46.3594 2.07317e-09       9.487729
#> 
#> $Conclusion
#> [1] "Reject"
#> 
meanTest.single(X, mu0, full = TRUE)
#> $Hypothesis
#> [1] "H0: mu = ( 5.8, 3, 4.3, 1.3 ) when Sigma is unknown"
#> 
#> $`Sample Mean`
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>       5.8433       3.0573       3.7580       1.1993 
#> 
#> $`Sample Diviation`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   102.168333   -6.322667     189.8730    76.92433
#> Sepal.Width     -6.322667   28.306933     -49.1188   -18.12427
#> Petal.Length   189.873000  -49.118800     464.3254   193.04580
#> Petal.Width     76.924333  -18.124267     193.0458    86.56993
#> 
#> $Statistics
#>     Statistics     Value p.value Critical.Value
#> 1 Hotelling T2 244.16277      NA             NA
#> 2            F  59.81168       0       2.433633
#> 
#> $`Degree of freedom`
#>     Statistics            df
#> 1 Hotelling T2 T2( 4 , 149 )
#> 2            F  F( 4 , 146 )
#> 
#> $Conclusion
#> [1] "Reject"
#>