Test of Two Mean Vectors

Hypothesis testing for equality of the mean vectors of two multivariate normal populations when the covariance matrices are equal or unequal.

Usage

meanTest.two(
  data1,
  data2,
  alpha = 0.05,
  equal = TRUE,
  method = c("None", "Coupled", "Transformed"),
  detail = TRUE
)

Arguments

data1

A matrix or data frame of group 1.

data2

A matrix or data frame of group 2.

alpha

The significance level. Default is 0.05.

equal

A boolean value. Default is TRUE. If TRUE, the covariance matrix is equal. If FALSE, the covariance matrix is not equal.

method

A string value. Default is "None". When equal is FALSE, you must choose a method in "Coupled" or "Transformed". Choose "Coupled" when the sample size of two groups is equal. Choose "Transformed" when the sample size of two groups is not equal.

detail

A boolean value. Default is TRUE. If TRUE, the detail of the test result will be displayed.

Value

A object of class "testResult", which is a list with the following elements: Return when the param equal is TRUE.

Stat

A data frame containing the statistics, p value and critical value.

SampMean1

The sample mean of group 1.

SampMean2

The sample mean of group 2.

SampA1

The sample deviation of group 1.

SampA2

The sample deviation of group 2.

Df

The degree of freedom.

Return when the param equal is FALSE and method is "Coupled".

Stat

A data frame containing the statistics, p value and critical value.

SampMeanC

The sample mean of coupled data.

SampAC

The sample deviation of coupled data.

Df

The degree of freedom.

dataC

The coupled data.

Return when the param equal is FALSE and method is "Transformed".

Stat

A data frame containing the statistics, p value and critical value.

SampMeanT

The sample mean of transformed data.

SampAT

The sample deviation of transformed data.

Df

The degree of freedom.

dataT

The transformed data. Return when the param equal is FALSE and method is "Transformed".

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.76-80.

Author

Xifeng Zhang

Examples

data(iris)
X <- iris[1:50, 1:4]
Y <- iris[51:100, 1:4]

test1 <- meanTest.two(X, Y)
#> H0: mu1 = mu2 when Sigma1 = Sigma2
#> The sample mean of group 1 is: (5.0063.4281.4620.246)'
#> The sample deviation of group 1 is:
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length       6.0882      4.8616       0.8014      0.5062
#> Sepal.Width        4.8616      7.0408       0.5732      0.4556
#> Petal.Length       0.8014      0.5732       1.4778      0.2974
#> Petal.Width        0.5062      0.4556       0.2974      0.5442
#> 
#> The sample mean of group 2 is: (5.9362.774.261.326)'
#> The sample deviation of group 2 is:
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      13.0552       4.174        8.962      2.7332
#> Sepal.Width        4.1740       4.825        4.050      2.0190
#> Petal.Length       8.9620       4.050       10.820      3.5820
#> Petal.Width        2.7332       2.019        3.582      1.9162
#> 
#> The statistics, p value and critical value are:
#>                  Value p.value Critical.Value
#> Hotelling T2 2580.8385      NA             NA
#> F             625.4583       0       2.467494
#> 
#> The degree of freedom is:
#>                      df
#> Hotelling T2 T2( 4 98 )
#> F             F( 4 95 )
#> 
#> Conclusion: Reject
test2 <- meanTest.two(X, Y, full = TRUE)
#> Error in meanTest.two(X, Y, full = TRUE): unused argument (full = TRUE)
test3 <- meanTest.two(X, Y, equal = FALSE, method = "Coupled")
#> Error in meanTest.two(X, Y, equal = FALSE, method = "Coupled"): object 'X.bar' not found
test4 <- meanTest.two(X, Y, equal = FALSE, method = "Transformed")
#> Error in meanTest.two(X, Y, equal = FALSE, method = "Transformed"): object 'A' not found

test1$Stat
#>                  Value p.value Critical.Value
#> Hotelling T2 2580.8385      NA             NA
#> F             625.4583       0       2.467494
test1$SampMean1
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.006        3.428        1.462        0.246 
test3$SampMeanC
#> Error: object 'test3' not found
test4$dataT
#> Error: object 'test4' not found