Test of Independence

Independence tests for multivariate normal populations.

Usage

indTest.multi(data, subdim = FALSE, alpha = 0.05, full = FALSE)

Arguments

data

The data matrix which is a matrix or data frame.

subdim

The dimensions of submatrices. The default is FALSE, which means the independence of all components of the random vector will be tested.

alpha

The significance level. Default is 0.05.

full

A boolean value. Default is FALSE. If TRUE, the full output will be displayed.

Value

If full is FALSE, a list containing the hypothesis, statistics and conclusion will be returned. If full is TRUE, a list containing the hypothesis, sample mean, sample diviation, sample diviation of submatrix, statistics, modify factor, degree of freedom and conclusion will be returned.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.92-94.

Author

Xifeng Zhang