Multiple Mean Vectors Hypothesis Testing

Test whether the mean vectors of multiple multivariate normal populations are all equal when the covariance matrices are equal. Suppose we have k populations, the null hypothesis is that "H0: mu1 = mu2 = ... = muk". There are two approximations (Bartlett's chi2 and Rao's F) to compute the p-value and the critical value. The realized value of the Wilks Lambda statistic and its degrees of freedom are also provided. If you want to perform an exact test, consult the Wilks Lambda statistic quantile table yourself, depending on the realized value of the statistic and its degrees of freedom.

Usage

meanTest.multi(X, label, alpha = 0.05, detail = TRUE)

Arguments

X

The data matrix which is a matrix or data frame.

label

A vector of group labels.

alpha

The significance level. Default is 0.05.

detail

A boolean value. Default is TRUE. If TRUE, the detail of the test result will be displayed. If FALSE, a simple summary will be displayed.

Value

An object of class "testResult", which is a list with the following elements:

Conclusion

The conclusion of the test.

Stat

A data frame containing the statistics, p value and critical value.

SampMeanT

The sample mean.

SampMeanWithin

The sample mean of each group.

SdTotal

The total sample deviation.

SdBetween

The sample deviation between group.

SdWithin

The sample deviation of each group.

SdWithinT

The sample deviation within group.

Df

The degree of freedom.

sampleSize

The sample size of each group.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.80-83.

Author

Xifeng Zhang

Examples

data(iris)
chart <- iris[, 1:4]
species <- iris[, 5]
# carry out the test
test1 <- meanTest.multi(chart, species)
#> H0: mu1 = mu2 = ... = muk, k = 3
#> The sample mean is: (5.84333.05733.7581.1993)'
#> The sample mean within group is:
#> $setosa
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.006        3.428        1.462        0.246 
#> 
#> $versicolor
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.936        2.770        4.260        1.326 
#> 
#> $virginica
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        6.588        2.974        5.552        2.026 
#> 
#> 
#> The total sample deviation is:
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   102.168333   -6.322667     189.8730    76.92433
#> Sepal.Width     -6.322667   28.306933     -49.1188   -18.12427
#> Petal.Length   189.873000  -49.118800     464.3254   193.04580
#> Petal.Width     76.924333  -18.124267     193.0458    86.56993
#> 
#> The sample deviation between group is:
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length     63.21213   -19.95267     165.2484    71.27933
#> Sepal.Width     -19.95267    11.34493     -57.2396   -22.93267
#> Petal.Length    165.24840   -57.23960     437.1028   186.77400
#> Petal.Width      71.27933   -22.93267     186.7740    80.41333
#> 
#> The sample deviation within group is:
#> $setosa
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length       6.0882      4.8616       0.8014      0.5062
#> Sepal.Width        4.8616      7.0408       0.5732      0.4556
#> Petal.Length       0.8014      0.5732       1.4778      0.2974
#> Petal.Width        0.5062      0.4556       0.2974      0.5442
#> 
#> $versicolor
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      13.0552       4.174        8.962      2.7332
#> Sepal.Width        4.1740       4.825        4.050      2.0190
#> Petal.Length       8.9620       4.050       10.820      3.5820
#> Petal.Width        2.7332       2.019        3.582      1.9162
#> 
#> $virginica
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      19.8128      4.5944      14.8612      2.4056
#> Sepal.Width        4.5944      5.0962       3.4976      2.3338
#> Petal.Length      14.8612      3.4976      14.9248      2.3924
#> Petal.Width        2.4056      2.3338       2.3924      3.6962
#> 
#> 
#> The sample deviation within group (total) is:
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      38.9562     13.6300      24.6246      5.6450
#> Sepal.Width       13.6300     16.9620       8.1208      4.8084
#> Petal.Length      24.6246      8.1208      27.2226      6.2718
#> Petal.Width        5.6450      4.8084       6.2718      6.1566
#> 
#> The statistics, p value and critical value are:
#>                        Value p.value   Critical.Value
#> Wilks Lambda    2.343863e-02                         
#> Bartlett's Chi2 5.461153e+02       0 15.5073130558655
#> Rao's F         1.274530e+04       0  1.9706194163219
#> 
#> The degree of freedom is:
#>                                    df
#> Wilks Lambda    Lambda( 4 , 147 , 2 )
#> Bartlett's Chi2             Chi2( 8 )
#> Rao's F                  F( 8 , 288 )
#> 
#> Conclusion: Bartlett's Chi-square:  Reject ; Rao's F:  Reject
test2 <- meanTest.multi(chart, species, detail = FALSE)
#> H0: mu1 = mu2 = ... = muk, k = 3
#> The statistics, p value and critical value are:
#>                        Value p.value   Critical.Value
#> Wilks Lambda    2.343863e-02                         
#> Bartlett's Chi2 5.461153e+02       0 15.5073130558655
#> Rao's F         1.274530e+04       0  1.9706194163219
#> 
#> Conclusion: Bartlett's Chi-square:  Reject ; Rao's F:  Reject
# get the elements
test1$Stat
#>                        Value p.value   Critical.Value
#> Wilks Lambda    2.343863e-02                         
#> Bartlett's Chi2 5.461153e+02       0 15.5073130558655
#> Rao's F         1.274530e+04       0  1.9706194163219
test1$SampMeanT
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>     5.843333     3.057333     3.758000     1.199333 
test1$sampleSize
#>     setosa versicolor  virginica 
#>         50         50         50