Test of Multiple Mean Vectors

Mean vector test for multiple multivariate normal totals when the covariance array is equal (multivariate analysis of variance).

meanTest.multi(X, label, alpha = 0.05, full = FALSE)

Arguments

X

The data matrix which is a matrix or data frame.

label

A vector of group labels.

alpha

The significance level. Default is 0.05.

full

A boolean value. Default is FALSE. If TRUE, the full output will be displayed.

Value

A list containing the hypothesis, sample size, total sample mean, within group mean, total sum of squares, within sum of squares, between sum of squares, statistics, df of Wilk's Lambda, df of Bartlett's Chi2, df of Rao's F, p value, approximate conclusion.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.80-83.

Examples

data(iris)
chart <- iris[, 1:4]
species <- iris[, 5]
meanTest.multi(chart, species)
#> $Hypothesis
#> [1] "H0: mu1 = mu2 = ... = muk, k =  3"
#> 
#> $Statistics
#>          Statistics        Value p.value Critical.Value
#>       Wilk's Lambda 2.343863e-02      NA             NA
#> df2 Bartlett's Chi2 5.461153e+02       0      15.507313
#> df1         Rao's F 1.274530e+04       0       1.970619
#> 
#> $`Approx Reject`
#> Bartlett      Rao 
#> "Reject" "Reject" 
#> 
meanTest.multi(chart, species, full = TRUE)
#> $Hypothesis
#> [1] "H0: mu1 = mu2 = ... = muk, k =  3"
#> 
#> $`Sample Size`
#>     setosa versicolor  virginica 
#>         50         50         50 
#> 
#> $`Total Sample Mean`
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>     5.843333     3.057333     3.758000     1.199333 
#> 
#> $`Within Group Mean`
#> $`Within Group Mean`$setosa
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.006        3.428        1.462        0.246 
#> 
#> $`Within Group Mean`$versicolor
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.936        2.770        4.260        1.326 
#> 
#> $`Within Group Mean`$virginica
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        6.588        2.974        5.552        2.026 
#> 
#> 
#> $`Total Sum of Squares`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length   102.168333   -6.322667     189.8730    76.92433
#> Sepal.Width     -6.322667   28.306933     -49.1188   -18.12427
#> Petal.Length   189.873000  -49.118800     464.3254   193.04580
#> Petal.Width     76.924333  -18.124267     193.0458    86.56993
#> 
#> $`Within Sum of Squares (Total)`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      38.9562     13.6300      24.6246      5.6450
#> Sepal.Width       13.6300     16.9620       8.1208      4.8084
#> Petal.Length      24.6246      8.1208      27.2226      6.2718
#> Petal.Width        5.6450      4.8084       6.2718      6.1566
#> 
#> $`Within Sum of Squares`
#> $`Within Sum of Squares`$setosa
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length       6.0882      4.8616       0.8014      0.5062
#> Sepal.Width        4.8616      7.0408       0.5732      0.4556
#> Petal.Length       0.8014      0.5732       1.4778      0.2974
#> Petal.Width        0.5062      0.4556       0.2974      0.5442
#> 
#> $`Within Sum of Squares`$versicolor
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      13.0552       4.174        8.962      2.7332
#> Sepal.Width        4.1740       4.825        4.050      2.0190
#> Petal.Length       8.9620       4.050       10.820      3.5820
#> Petal.Width        2.7332       2.019        3.582      1.9162
#> 
#> $`Within Sum of Squares`$virginica
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      19.8128      4.5944      14.8612      2.4056
#> Sepal.Width        4.5944      5.0962       3.4976      2.3338
#> Petal.Length      14.8612      3.4976      14.9248      2.3924
#> Petal.Width        2.4056      2.3338       2.3924      3.6962
#> 
#> 
#> $`Between Sum of Squares`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length     63.21213   -19.95267     165.2484    71.27933
#> Sepal.Width     -19.95267    11.34493     -57.2396   -22.93267
#> Petal.Length    165.24840   -57.23960     437.1028   186.77400
#> Petal.Width      71.27933   -22.93267     186.7740    80.41333
#> 
#> $Statistics
#>          Statistics        Value p.value Critical.Value
#>       Wilk's Lambda 2.343863e-02      NA             NA
#> df2 Bartlett's Chi2 5.461153e+02       0      15.507313
#> df1         Rao's F 1.274530e+04       0       1.970619
#> 
#> $`Degree of freedom`
#>        Statistics                    df
#> 1   Wilk's Lambda Lambda( 4 , 147 , 2 )
#> 2 Bartlett's Chi2             Chi2( 8 )
#> 3         Rao's F          F( 8 , 288 )
#> 
#> $`Approx Reject`
#> Bartlett      Rao 
#> "Reject" "Reject" 
#>