Test of Two Mean Vectors

Hypothesis testing for equality of the mean vectors of two multivariate normal populations when the covariance matrices are equal or unequal.

meanTest.two(
  data1,
  data2,
  alpha = 0.05,
  equal = TRUE,
  method = c("None", "Coupled", "Transformed"),
  full = FALSE
)

Arguments

data1

A matrix or data frame of group 1.

data2

A matrix or data frame of group 2.

alpha

The significance level. Default is 0.05.

equal

A boolean value. Default is TRUE. If TRUE, the covariance matrix is equal. If FALSE, the covariance matrix is not equal.

method

A string value. Default is "None". When equal is FALSE, you must choose a method in "Coupled" or "Transformed". Choose "Coupled" when the sample size of two groups is equal. Choose "Transformed" when the sample size of two groups is not equal. If you want to use the likelihood ratio test, please use the function "mean.test.multi".

full

A boolean value. Default is FALSE. If TRUE, the full output will be displayed.

Value

A list containing the hypothesis, sample mean, sample diviation, statistics, df of T2, df of F, p value, critical value and conclusion.

References

Huixuan, Gao. Applied Multivariate Statistical Analysis. Peking University Press, 2005: pp.76-80.

Examples

data(iris)
X <- iris[1:50, 1:4]
Y <- iris[51:100, 1:4]
meanTest.two(X, Y)
#> $Hypothesis
#> [1] "H0: mu1 = mu2, with unknown but equal covariance matrix"
#> 
#> $Statistics
#>     Statistics     Value p.value Critical.Value
#> 1 Hotelling T2 2580.8385      NA             NA
#> 2            F  625.4583       0       2.467494
#> 
#> $Conclusion
#>      [,1]    
#> [1,] "Reject"
#> 
meanTest.two(X, Y, full = TRUE)
#> $Hypothesis
#> [1] "H0: mu1 = mu2, with unknown but equal covariance matrix"
#> 
#> $`Sample Mean X`
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.006        3.428        1.462        0.246 
#> 
#> $`Sample Mean Y`
#> Sepal.Length  Sepal.Width Petal.Length  Petal.Width 
#>        5.936        2.770        4.260        1.326 
#> 
#> $`Sample Diviation A1`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length       6.0882      4.8616       0.8014      0.5062
#> Sepal.Width        4.8616      7.0408       0.5732      0.4556
#> Petal.Length       0.8014      0.5732       1.4778      0.2974
#> Petal.Width        0.5062      0.4556       0.2974      0.5442
#> 
#> $`Sample Diviation A2`
#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length      13.0552       4.174        8.962      2.7332
#> Sepal.Width        4.1740       4.825        4.050      2.0190
#> Petal.Length       8.9620       4.050       10.820      3.5820
#> Petal.Width        2.7332       2.019        3.582      1.9162
#> 
#> $Statistics
#>     Statistics     Value p.value Critical.Value
#> 1 Hotelling T2 2580.8385      NA             NA
#> 2            F  625.4583       0       2.467494
#> 
#> $`Degree of freedom`
#>     Statistics         df
#> 1 Hotelling T2 T2( 4 98 )
#> 2            F  F( 4 95 )
#> 
#> $Conclusion
#>      [,1]    
#> [1,] "Reject"
#> 
meanTest.two(X, Y, equal = FALSE, method = "Coupled")
#> $Hypothesis
#> [1] "H0: mu1 = mu2, with unknown but different covariance matrix"
#> 
#> $Statistics
#>     Statistics    Value      p.value Critical.Value
#> 1 Hotelling T2 45.86582           NA             NA
#> 2            F 10.76443 3.051338e-06       2.574035
#> 
#> $Conclusion
#>      [,1]    
#> [1,] "Reject"
#> 
meanTest.two(X, Y, equal = FALSE, method = "Transformed")
#> $Hypothesis
#> [1] "H0: mu1 = mu2, with unknown but different covariance matrix"
#> 
#> $Statistics
#>     Statistics    Value      p.value Critical.Value
#> 1 Hotelling T2 45.86582           NA             NA
#> 2            F 10.76443 3.051338e-06       2.574035
#> 
#> $Conclusion
#>      [,1]    
#> [1,] "Reject"
#>